+++ bin/test-node.sh
@@ -20,6 +20,9 @@ fi
 if [ $TYPE = "mapreduce" ]; then
     TESTS_PATH="tests/mapreduce/test.*.js"
 fi
+if [ $TYPE = "find" ]; then
+    TESTS_PATH="tests/find/*/test.*.js"
+fi
 if [ $COVERAGE ]; then
     # run all tests when testing for coverage
     TESTS_PATH="tests/{unit,integration,mapreduce,component}/test*.js"
