+++ tests/find/index.html
@@ -19,7 +19,7 @@
   <body>
     <div id="mocha"></div>
     <script src='../integration/utils-bundle.js'></script>
-    <script src='./test-setup.js'></script>
+    <script src='./test-suite-1/test.aaa-setup.js'></script>
     <script src='./test-suite-1/test.and.js'></script>
     <script src='./test-suite-1/test.array.js'></script>
     <script src='./test-suite-1/test.basic.js'></script>
@@ -50,6 +50,8 @@
     <script src='./test-suite-1/test.sorting.js'></script>
     <script src='./test-suite-1/test.type.js'></script>
     <script src='./test-suite-1/test.zzz-suite-1.js'></script>
+    <script src='./test-suite-2/test.kitchen-sink.js'></script>
+    <script src='./test-suite-2/test.kitchen-sink-2.js'></script>
     <script type="text/javascript" src="../integration/webrunner.js"></script>
   </body>
 </html>
