+++ tests/find/test-suite-1/test.aaa-setup.js
@@ -1,9 +1,6 @@
-// to avoid having to bundle anything, we just use a global window.testCases
-// object and then run those
-
-'use strict';
-
 if (typeof process === 'undefined' || process.browser) {
+  // to avoid having to bundle anything, we just use a global window.testCases
+// object and then run those
   window.testCases = [];
 } else { // node
   global.testCases = [];
